requirements.txt
Flask==2.2.5
mysql-connector-python==8.1.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.24.1
Werkzeug==2.2.3
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
